Start working on item portal improvements
