+++ packages/react-native-sortables/src/providers/shared/hooks/useItemDecorationStyles.ts
@@ -15,24 +15,6 @@ import type { AnimatedStyleProp } from '../../../types';
 import { ItemPortalState } from '../../../types';
 import { useCommonValuesContext } from '../CommonValuesProvider';
 
-const TELEPORTED_ITEM_STYLE: ViewStyle = {
-  maxHeight: 0,
-  opacity: 0,
-  ...Platform.select({
-    android: {
-      elevation: 0
-    },
-    default: {},
-    native: {
-      shadowOpacity: 0
-    }
-  })
-};
-
-const NOT_TELEPORTED_ITEM_STYLE: ViewStyle = {
-  maxHeight: 9999 // 'auto' doesn't trigger onLayout on web/android
-};
-
 export default function useItemDecorationStyles(
   itemKey: string,
   isActive: SharedValue<boolean>,
@@ -67,10 +49,6 @@ export default function useItemDecorationStyles(
   });
 
   const animatedStyle = useAnimatedStyle(() => {
-    if (portalState?.value === ItemPortalState.TELEPORTED) {
-      return TELEPORTED_ITEM_STYLE;
-    }
-
     const progress = activationAnimationProgress.value;
     const zeroProgressOpacity = interpolate(
       adjustedInactiveProgress.value,
@@ -94,7 +72,6 @@ export default function useItemDecorationStyles(
       : { shadowColor };
 
     return {
-      ...NOT_TELEPORTED_ITEM_STYLE,
       ...shadow,
       opacity: interpolate(
         progress,
